Add database connection file
Because we will have more than one file that work with the
database and almost all of them first need to make a
connection with the database, so instead of duplicating
the database connection boilerplate code in every single
file, we can simply wrap it into a connection file and just
include it in those other files. It also makes the code
more maintainable and easier to debug.