Get a four-item list of products (Lazy loading)
Instead of loading the entire table and rendering it to the
user in one go as in bulk loading, the concept of lazy
loading assists in loading only the required section and
delays the remaining, until it is needed by the user. Lazy
loading causes improved performance and Less traffic load
for the host.